+++ ubuntu14/nginx-php5-fpm.sh
@@ -20,10 +20,7 @@ declare -r VHOST_PHP5_FPM_SAV=/etc/nginx/sites-available/${VHOST_PHP5_FPM_CONF}
 : "----- put test php file on document root"
 declare -r DOC_ROOT=/usr/share/nginx/html
 declare -r VHOST_DOC_ROOT=${DOC_ROOT}/${VHOST_PHP5_FPM}
-if [ ! -d ${VHOST_DOC_ROOT} ]
-then
-  mkdir -p ${VHOST_DOC_ROOT}
-fi
+[ -d ${VHOST_DOC_ROOT} ] || mkdir -p ${VHOST_DOC_ROOT}
 
 declare -ar COPY_TARGETS=(${VHOST_PHP5_FPM_SAV} ${VHOST_DOC_ROOT}/index.php)
 for target in ${COPY_TARGETS[@]}
@@ -36,10 +33,7 @@ sed -i "s|@@FCGI_PASS@@|${FCGI_PASS}|g" ${VHOST_PHP5_FPM_SAV}
 
 : "----- symlink vhost conf to sites-enabled"
 declare -r VHOST_PHP5_FPM_ENB=/etc/nginx/sites-enabled/${VHOST_PHP5_FPM_CONF}
-if [ -L ${VHOST_PHP5_FPM_ENB} ]
-then
-  rm ${VHOST_PHP5_FPM_ENB}
-fi
+[ -L ${VHOST_PHP5_FPM_ENB} ] && rm ${VHOST_PHP5_FPM_ENB}
 ln -s ${VHOST_PHP5_FPM_SAV} ${VHOST_PHP5_FPM_ENB}
 
 /etc/init.d/nginx restart
