+++ ubuntu14/php5-fpm.sh
@@ -13,10 +13,8 @@ ${PRVENV_CMD_PKG_INS} php5-cli php5-curl php5-gd php5-memcache php5-memcached ph
 
 : "----- modify php.ini for fpm"
 declare -r PHP_INI_FILE=/etc/php5/fpm/php.ini
-if [ ! -f ${PHP_INI_FILE}.org ]
-then
-  cp ${PHP_INI_FILE} ${PHP_INI_FILE}.org
-fi
+[ -f ${PHP_INI_FILE}.org ] || cp ${PHP_INI_FILE} ${PHP_INI_FILE}.org
+
 ## fix_pathinfo=1 is insecure
 sed -i "s/;cgi.fix_pathinfo=1/cgi.fix_pathinfo=0/g" ${PHP_INI_FILE}
